requirements.txt
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.6
google-analytics-data==0.16.2